Added setter and getters for cellID

+++ include/Pixel.h
@@ -5,7 +5,7 @@
 class Pixel
 {
     private:
-    int cellID;
+    long int cellID;
     public:
         /** Default constructor */
         Pixel();
